server/requirements.txt
Django>=5.2,<6.0
python-socketio>=5.13,<6.0
uvicorn>=0.35,<1.0
